% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_Luminex_csv.R
\name{read_Xponent_csv}
\alias{read_Xponent_csv}
\title{Read a Luminex-Xponent generated csv file}
\usage{
read_Xponent_csv(path)
}
\arguments{
\item{path}{The path to a Luminx csv file}
}
\value{
A list of lists with all the information from the file.
    At the top level there are two lists
    \itemize{
        \item \code{BatchHeader} that stores all the information before the \emph{Results}
        section as well as the CRC and
        \item \code{AssayData} that stores all the information in the \emph{Results} section
    }

    \code{BatchHeader} stores data as key-value pairs with keys taken from the first column of the csv file and
    values from the rest. Values are mostly character-variables and there are few character vectors.
    The only exception is \code{CALInfo} with is a list of dataframes containing information about:
    \itemize{
        \item Last Calibration in \code{LastCalibration}
        \item Classification Calibrator in \code{ClassificationCalibratorExtended}
        \item Reporter Calibratorin (you guessed it!) \code{ReporterCalibrator}
    }

    \code{AssayData} stores data as dataframes, except for \code{Run} data which is a list of dataframes
    (Audit Logs and Warning/Errors). There are 4 dataframes in \code{AssayData}:
    \itemize{
        \item \strong{Exprs} Expression data, ie \emph{raw} data, including \strong{Medians, Counts, Net MFI}, etc
        \item \strong{Average} Averages for replicate wells
        \item \strong{Well} Information about each well (currently only the \emph{Dilution Factor})
        \item \strong{Bead} Information about the beads, eg BeadID
    }
    more information about the different data generated can be found
    \href{https://www.luminexcorp.com/blog/2012/06/19/its-all-about-the-stats/}{here}.
}
\description{
Read a Luminex-Xponent generated csv file
}
\details{
The format details (eg comma or semicolon delimiter) are infered by the function.
    There is a known edge-case, for some semicolon-separated files,
    where some calibration information is misplaced. This is the **only** case where you should
    mess with the csv file, you should **not mess** with it.
}
\examples{
## Read a csv file for a 384 plate
path <- system.file("extdata", "test0_384.csv", package="moach")
lumcsv <- read_Xponent_csv(path)
}
\author{
Teo Sakel
}

